Adding exponential backoff to polling (#241)
I thought we had this in place previously, but I guess not. The idea here is that unless a DAG specifies a polling interval, we can back off from the 15 second default polling interval the longer that a task is running. In this case, we will start at 15 seconds, going up by 25% each cycle, to a maximum of 300 seconds (5 minutes). This should take significant load off of the Kuberntes controller.

https://bluecore.atlassian.net/browse/PROD-27345